+++ src/model/shibboleth_config.rs
@@ -11,6 +11,7 @@ pub struct ShibbolethConfig {
     pub attribute_filter_paths: Vec<String>,
     pub application_defaults: Option<ApplicationDefaults>,
     pub status_handler: Option<StatusHandler>,
+    pub errors: Option<ErrorsConfig>,
 }
 
 #[derive(Debug, Default)]
@@ -25,8 +26,12 @@ pub struct SessionsConfig {
     pub handler_url: Option<String>,
     pub handler_ssl: Option<String>,
     pub cookie_props: Option<String>,
+    pub lifetime: Option<String>,
+    pub timeout: Option<String>,
     pub has_sso: bool,
     pub has_session_initiator: bool,
+    pub has_logout: bool,
+    pub sso_entity_id: Option<String>,
 }
 
 #[derive(Debug)]
@@ -41,6 +46,7 @@ pub struct MetadataProvider {
 }
 
 #[derive(Debug)]
+#[allow(dead_code)]
 pub struct MetadataFilter {
     pub filter_type: String,
     pub certificate: Option<String>,
@@ -49,6 +55,7 @@ pub struct MetadataFilter {
 }
 
 #[derive(Debug)]
+#[allow(dead_code)]
 pub struct CredentialResolver {
     pub resolver_type: String,
     pub use_attr: Option<String>,
@@ -59,4 +66,18 @@ pub struct CredentialResolver {
 #[derive(Debug, Default)]
 pub struct StatusHandler {
     pub acl: Option<String>,
+}
+
+#[derive(Debug, Default)]
+#[allow(dead_code)]
+pub struct ErrorsConfig {
+    pub support_contact: Option<String>,
+    pub help_location: Option<String>,
+    pub style_sheet: Option<String>,
+    pub session_error: Option<String>,
+    pub access_error: Option<String>,
+    pub ssl_error: Option<String>,
+    pub local_logout: Option<String>,
+    pub metadata_error: Option<String>,
+    pub global_logout: Option<String>,
 }